+++ lib/src/invitation/service.dart
@@ -248,7 +248,7 @@ class ZegoUIKitPrebuiltCallInvitationService
 
     await ZegoUIKit().getZegoUIKitVersion().then((uikitVersion) {
       ZegoLoggerService.logInfo(
-        'versions: zego_uikit_prebuilt_call:4.17.0; $uikitVersion',
+        'versions: zego_uikit_prebuilt_call:4.17.0-beta.1; $uikitVersion',
         tag: 'call-invitation',
         subTag: 'service(${identityHashCode(this)}), init',
       );